refactor(persistence): ADR-0035 — relocate RegisteredNode to kirra-core (persistence enabling slice 1)
A dependency audit of verifier_store after the seam step found the ADR's
"mechanical extraction" claim was optimistic: the module still names several
root-crate items (crate::audit_chain::*, crate::verifier::RegisteredNode,
audit_shipper::ShippedAuditRecord, verdicts::is_valid_verdict_id,
key_registry::KeyRegistry) that would cycle if verifier_store moved to its own
crate. kirra-persistence needs a few enabling slices first.

This is enabling slice 1 — the cleanest: RegisteredNode is a plain data struct
(no impls; only non-std field is NodeTrustState, already in kirra-core), so it
moves to kirra-core alongside NodeTrustState and is re-exported as
crate::verifier::RegisteredNode. Every existing crate::verifier::RegisteredNode /
kirra_verifier::RegisteredNode path (store, service bin, kirra-verifier-pg,
tests) resolves unchanged. No new crate, no Cargo.toml change → no lockfile churn.

ADR-0035 updated: corrected the "mechanical" claim with the real remaining-coupling
inventory and the enabling-slice sequence (RegisteredNode done; the audit_chain
pure-core split is the crux next slice).

Behaviour-preserving. Verified: kirra-core (184) + `cargo test --lib` (760) +
power-loss drill (3) + fmt + guardrails + orphan-cores green.
